Added LatestDraws view.

+++ backend/src/main/java/com/lottery/entities/LotteryDraw.java
@@ -15,7 +15,7 @@ public class LotteryDraw {
     @GeneratedValue
     private Long id;
     private Timestamp drawTime;
-    private int numbers;
+    private long numbers;
     @OneToMany(mappedBy = "lotteryDraw")
     private List<Entry> entries;
     @OneToMany(mappedBy = "lotteryDraw")
